refactor(backgroundjobs): relocate Hangfire adapters to API/Infrastructure
Move HangfireJobEnqueuer and HangfireRecurringJobScheduler from
Anela.Heblo.Application/Features/BackgroundJobs/Services/ to
Anela.Heblo.API/Infrastructure/Hangfire/. The interfaces (IHangfireJobEnqueuer,
IHangfireRecurringJobScheduler) remain in Application so handlers continue to
depend only on Application-layer abstractions.

DI registration moves from BackgroundJobsModule.AddBackgroundJobsModule to
ServiceCollectionExtensions.AddHangfireServices, co-located with the rest of
the Hangfire wiring. Lifetimes are preserved (Scoped enqueuer, Singleton scheduler).

Note: Hangfire.Core PackageReference is intentionally kept in
Anela.Heblo.Application.csproj — six other files in Application still import
Hangfire (GenerateArticleHandler, GenerateArticleJob, PlaudPollingJob,
ProductExportDownloadJob, DashboardModule, FailedJobsTile). Removing the
PackageReference is tracked as follow-up work.
