+++ backend/src/Anela.Heblo.API/Extensions/ServiceCollectionExtensions.cs
@@ -21,6 +21,7 @@ using Anela.Heblo.Adapters.Cups;
 using Anela.Heblo.Adapters.Cups.Features.ExpeditionList;
 using Anela.Heblo.API.Features.ExpeditionList;
 using Anela.Heblo.API.PDFPrints;
+using Anela.Heblo.Application.Features.BackgroundJobs.Services;
 using Anela.Heblo.Application.Features.ExpeditionList.Services;
 using Anela.Heblo.Application.Shared.Printing;
 using Anela.Heblo.Application.Features.Manufacture.UseCases.GetManufactureProtocol;
@@ -337,6 +338,12 @@ public static class ServiceCollectionExtensions
 
         // Register IBackgroundWorker implementation
         services.AddTransient<IBackgroundWorker, HangfireBackgroundWorker>();
+
+        // Register Hangfire adapter implementations (interfaces live in Application,
+        // concrete types live in API/Infrastructure/Hangfire — relocated to keep the
+        // Application project free of Hangfire imports for these specific adapters).
+        services.AddScoped<IHangfireJobEnqueuer, HangfireJobEnqueuer>();
+        services.AddSingleton<IHangfireRecurringJobScheduler, HangfireRecurringJobScheduler>();
 
         // Note: IRecurringJobStatusChecker is now registered in Application layer (BackgroundJobsModule)
