+++ backend/src/Anela.Heblo.Application/Features/BackgroundJobs/BackgroundJobsModule.cs
@@ -1,4 +1,3 @@
-using Anela.Heblo.Application.Features.BackgroundJobs.Services;
 using Anela.Heblo.Domain.Features.BackgroundJobs;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -10,13 +9,12 @@ public static class BackgroundJobsModule
     {
         // MediatR handlers are automatically registered by MediatR scan
         // Repository is registered in PersistenceModule
+        // Hangfire adapter implementations (IHangfireJobEnqueuer, IHangfireRecurringJobScheduler)
+        // are registered in Anela.Heblo.API.Extensions.ServiceCollectionExtensions.AddHangfireServices
+        // because their implementations live in the API project (Clean Architecture dependency rule).
 
         // Register recurring job status checker
         services.AddScoped<IRecurringJobStatusChecker, RecurringJobStatusChecker>();
-
-        // Register Hangfire job enqueuer
-        services.AddScoped<IHangfireJobEnqueuer, HangfireJobEnqueuer>();
-        services.AddSingleton<IHangfireRecurringJobScheduler, HangfireRecurringJobScheduler>();
 
         return services;
     }
